Refactor - db enum types into enums crate
