+++ crates/_types/enums/src/lib.rs
@@ -28,3 +28,5 @@ mod macros;
 
 pub mod core;
 pub mod files;
+pub mod ml;
+pub mod workers;
