+++ crates/_types/enums/src/ml/vocoder_type.rs
@@ -33,7 +33,7 @@ impl VocoderType {
 
 #[cfg(test)]
 mod tests {
-  use crate::db::enums::vocoder_type::VocoderType;
+  use crate::ml::vocoder_type::VocoderType;
   use crate::test_helpers::assert_serialization;
 
   #[test]
