+++ crates/_types/enums/src/workers/mod.rs
@@ -1,0 +1,2 @@
+pub mod generic_inference_type;
+pub mod generic_download_type;
