+++ crates/_types/reusable_types/src/db/mod.rs
@@ -1,2 +1,1 @@
-pub mod enums;
 pub mod payloads;
