+++ crates/lib/database_queries/src/queries/tts/tts_models/get_tts_model_for_inference.rs
@@ -4,10 +4,10 @@
 #![forbid(unused_variables)]
 
 use chrono::{Utc, DateTime};
+use enums::ml::vocoder_type::VocoderType;
 use log::warn;
 use sqlx::MySqlPool;
 use sqlx;
-use reusable_types::db::enums::vocoder_type::VocoderType;
 
 // TODO: Can probably just reuse another query.
 
@@ -91,7 +91,7 @@ SELECT
     tts.maybe_default_pretrained_vocoder,
 
     tts.maybe_custom_vocoder_token,
-    vocoder.vocoder_type as `maybe_custom_vocoder_type: reusable_types::db::enums::vocoder_type::VocoderType`,
+    vocoder.vocoder_type as `maybe_custom_vocoder_type: enums::ml::vocoder_type::VocoderType`,
     vocoder.title as maybe_custom_vocoder_title,
     vocoder.private_bucket_hash as maybe_custom_vocoder_private_bucket_hash,
