+++ crates/lib/database_queries/src/queries/vocoder/insert_vocoder_model.rs
@@ -7,7 +7,7 @@ use anyhow::anyhow;
 use container_common::anyhow_result::AnyhowResult;
 use crate::tokens::Tokens;
 use enums::core::visibility::Visibility;
-use reusable_types::db::enums::vocoder_type::VocoderType;
+use enums::ml::vocoder_type::VocoderType;
 use sqlx::MySqlPool;
 use std::path::Path;
