+++ crates/service/web/storyteller_web/src/http_server/endpoints/vocoders/get_vocoder.rs
@@ -12,8 +12,8 @@ use crate::http_server::web_utils::response_error_helpers::to_simple_json_error;
 use crate::server_state::ServerState;
 use database_queries::queries::vocoder::get_vocoder_model::get_vocoder_model_by_token;
 use enums::core::visibility::Visibility;
+use enums::ml::vocoder_type::VocoderType;
 use log::warn;
-use reusable_types::db::enums::vocoder_type::VocoderType;
 use std::fmt;
 use std::sync::Arc;
